A-CodeQuality
Parser and Command class had methods that were too long. Abstracting
out the logic for each of the command types and creating new classes for
each of them improves the overall quality of the code.
