+++ src/main/java/duke/Duke.java
@@ -32,8 +32,7 @@ public class Duke {
     }
 
     /**
-     * You should have your own function to generate a response to user input.
-     * Replace this stub with your completed method.
+     * Get duke response after receiving input from the User.
      */
     public String getResponse(String input) {
         try {
